etl/requirements.txt
pymongo
vaderSentiment
SQLAlchemy
psycopg2-binary
pandas